Question2/requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
matplotlib==3.9.2
tqdm==4.66.5
scikit-learn==1.5.1
gdown==5.2.0
streamlit==1.39.0
Pillow==10.4.0
